# MRG: download via NCBI REST API