--- a/src/lib.rs
+++ b/src/lib.rs
@@ -49,7 +49,7 @@ fn set_tokio_thread_pool(num_threads: usize) -> PyResult<usize> {
 
 #[pyfunction]
 #[allow(clippy::too_many_arguments)]
-#[pyo3(signature = (input_csv, param_str, failed_csv, failed_checksums, retry_times, fasta_location, keep_fastas, genomes_only, proteomes_only, download_only, batch_size, n_permits, output_sigs=None))]
+#[pyo3(signature = (input_csv, param_str, failed_csv, failed_checksums, retry_times, fasta_location, keep_fastas, genomes_only, proteomes_only, download_only, batch_size, n_permits, api_key, output_sigs=None))]
 fn do_gbsketch(
     py: Python,
     input_csv: String,
@@ -64,6 +64,7 @@ fn do_gbsketch(
     download_only: bool,
     batch_size: u32,
     n_permits: usize,
+    api_key: String,
     output_sigs: Option<String>,
 ) -> anyhow::Result<u8> {
     match directsketch::gbsketch(
@@ -80,6 +81,7 @@ fn do_gbsketch(
         download_only,
         batch_size,
         n_permits,
+        api_key,
         output_sigs,
     ) {
         Ok(_) => Ok(0),