For `gbsketch`, if we can download via NCBI REST API, we can up the number of simultaneous downloads. ref #180 

Major changes:
- ftp_path is no longer used, so it is removed from the input file
- change default retry times to 3 for both commands


Minor difference for `gbsketch` retries -- we no longer retry upon failed md5sum (because we download prot + dna at the same time, logic gets a bit more complex to make sure we can still record which one had md5sum failures without failing at any others without md5sum issues). I would expect most issues to arise from failures reading the zipfile, not failed checksums, so I don't anticipate issues. But happy to revisit if things come up.

Note: processing the zipfile with `async_zip` was failing, so i decided to try processing with regular zip. I'm not sure the performance implications of this.

remaining to do:
- [x] add back retries
- [x] docs about API key, n simultaneous downloads
- [x] clean up unused functions
- [x] clean up print statements
- [x] remove `ftp_path` column as it is no longer used. Update README + tests/test-data to match
- [x] Do we need to use virus-specific GET for viral genomes? No.
- [x] try using api key
- [x] benchmark main vs this PR. Should be able to use `-n 10` for this PR even with large genomes.
 ---> note, should we just set `-n 9` if user provides API KEY? Or maybe just recommend it in docs.