@bluegenes 2025-01-16
basic benchmarking with first 100 fungal accessions from https://github.com/ctb/2025-ncbi-rest-api:

## This branch:
this branch with `-n 9 -r 3`: 
```
time sourmash scripts gbsketch outputs/fungi-links.head100.csv -n 9 -c 8 -p k=21,k=31,k=51,dna  --failed fungi.gbsketch-fail.181.txt -r 3 --checksum-fail fungi.gbsketch-check-fail.181.txt -o fungi.head100.181.zip --api-key API_KEY
```
185.33s user 4.65s system 368% cpu **51.597 total**


this branch with `-n 10 -r 3` (`-n 10 -r 1` was dropping ~one sig each run):
```
 time sourmash scripts gbsketch outputs/fungi-links.head100.csv -n 10 -c 8 -p k=21,k=31,k=51,dna --failed fungi.gbsketch-fail.181.n10r3.txt -r 3 --checksum-fail fungi.gbsketch-check-fail.181.n10r3.txt -o fungi.head100.181.n10r3.zip --api-key API_KEY
```
194.04s user 4.92s system 403% cpu **49.320 total**

## main
main with `-n 2 -r 3`:
```
time sourmash scripts gbsketch outputs/fungi-links.head100.main.csv -n 2 -c 8 -p k=21,k=31,k=51,dna --failed fungi.gbsketch-fail.main.1.txt -r 1 --checksum-fail fungi.gbsketch-check-fail.main.1.txt -o fungi.head100.main.1.zip
```
157.94s user 3.59s system 104% cpu **2:34.14 total**


main with `-n 3 -r 3`:
```
time sourmash scripts gbsketch outputs/fungi-links.head100.main.csv -n 3 -c 8 -p k=21,k=31,k=51,dna --failed fungi.gbsketch-fail.main.n3r3.txt -r 3 --checksum-fail fungi.gbsketch-check-fail.main.n3r3.txt -o fungi.head100.main.n3r3.zip
```
157.27s user 3.10s system 143% cpu **1:51.65 total**

All of these commands completed without failing on any sketches.