@ctb 2025-01-16
w00t! two questions -

first, did you want me to review without really trying code so we can merge, and then adjust later? or would you prefer to give this a real workout over the next few days?

second, I don't understand the note:

> NOTE: -n 10 -r 1 was dropping ~one sig each run -- need either -n 9 OR -r 3 (or at least r>1). This is similar for main -- either run with -n 2 or up the number of retries

Does this mean we get incomplete results with specific args? If so, could we add an assertion or a failure or something when the args are known to be wrong? Otherwise I will forget :)