@bluegenes 2025-01-16
> w00t! two questions -
> 
> 
> 
> first, did you want me to review without really trying code so we can merge, and then adjust later? or would you prefer to give this a real workout over the next few days?

Give it a real workout please.

> second, I don't understand the note:
> 
> 
> 
> > NOTE: -n 10 -r 1 was dropping ~one sig each run -- need either -n 9 OR -r 3 (or at least r>1). This is similar for main -- either run with -n 2 or up the number of retries
> 
> 
> 
> Does this mean we get incomplete results with specific args? If so, could we add an assertion or a failure or something when the args are known to be wrong? Otherwise I will forget :)


-n is the number of simultaneous downloads, and -r is the number of retry times. What I'm saying here is that if we use the maximum number of simultaneous downloads stated by NCBI for either method, if we don't allow sufficient retry times, we will get some random missed signatures due to failed downloads.

This problem disappears with using one fewer n_simultaneous download, or by upping the number of retries from 1 to 3.

In this PR, I changed the default retries to 3, but I am also kind of leaning towards just suggesting using one fewer simultaneous download (e.g -n 9) so that fewer retries will be needed.

Alternatively, we could add a short time delay to the retry - not sure if it would help the first retry be more likely to be successful.